backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "docling>=2.63.0",
    "fastapi[standard]>=0.121.3",
    "dotenv>=0.9.9",
    "google-genai>=1.51.0",
    "pandas>=2.3.3",
    "pillow>=11.3.0",
    "pytesseract>=0.3.13",
    "scikit-image>=0.25.2",
    "ultralytics>=8.3.229",
    "easyocr>=1.7.2",
]
